add customer service

+++ src/main/java/ma/uca/ensas/ecommercebackendspring/service/DeliveryManService.java
@@ -2,6 +2,7 @@ package ma.uca.ensas.ecommercebackendspring.service;
 
 import lombok.RequiredArgsConstructor;
 import ma.uca.ensas.ecommercebackendspring.entities.DeliveryMan;
+import ma.uca.ensas.ecommercebackendspring.exceptions.ApiRequestException;
 import ma.uca.ensas.ecommercebackendspring.repositories.DeliveryManRepository;
 import org.springframework.stereotype.Service;
 
@@ -11,7 +12,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class DeliveryManService {
 
-    private DeliveryManRepository deliveryManRepository;
+    private final DeliveryManRepository deliveryManRepository;
 
     public DeliveryMan saveDeliveryMan(DeliveryMan deliveryMan){
         return deliveryManRepository.save(deliveryMan);
@@ -22,6 +23,9 @@ public class DeliveryManService {
     }
 
     public void deleteDeliveryManById(Long id){
+        if (deliveryManRepository.findById(id).isEmpty()){
+            throw new ApiRequestException("Delivery man with id : " + id + " doesn't exists");
+        }
         deliveryManRepository.deleteById(id);
     }
 }
